% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vibe_palette}
\alias{vibe_palette}
\title{Get a Vibe Color Palette}
\usage{
vibe_palette(name = "Default", n = NULL, reverse = FALSE)
}
\arguments{
\item{name}{The name of the palette to retrieve. Use \code{names(vibe_palettes)}
to see available options. Default is "Default".}

\item{n}{The number of colors to return. If \code{n} is less than or equal to the
number of colors in the palette, the first \code{n} colors are returned.
If \code{n} is greater, colors will be interpolated across the palette range.}

\item{reverse}{Logical. Should the palette be reversed? Default is \code{FALSE}.}
}
\value{
A character vector of hex color codes.
}
\description{
Retrieve a color palette by name. Returns the hex color codes
for the specified palette.
}
\examples{
# Get a palette (when palettes are defined)
# vibe_palette()
# vibe_palette("Crimson")
# vibe_palette("Crimson", n = 10)
}
