% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibeColors-package.R
\docType{package}
\name{vibeColors-package}
\alias{vibeColors}
\alias{vibeColors-package}
\title{vibeColors: Color Palettes for ggplot2}
\description{
A collection of color palettes designed for use with ggplot2.
Provides discrete and continuous color scales that integrate seamlessly
with the ggplot2 plotting system.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/yourusername/vibeColors/issues}
}
}
\author{
\strong{Maintainer}: Your Name \email{your.email@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}