% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{vibe_palette}
\alias{vibe_palette}
\title{Get a Vibe Color Palette}
\usage{
vibe_palette(name, n = NULL, reverse = FALSE)
}
\arguments{
\item{name}{The name of the palette to retrieve. Use \code{names(vibe_palettes)}
to see available options.}

\item{n}{The number of colors to return. If \code{n} is greater than the number
of colors in the palette, colors will be interpolated.}

\item{reverse}{Logical. Should the palette be reversed? Default is \code{FALSE}.}
}
\value{
A character vector of hex color codes.
}
\description{
Retrieve a color palette by name. Returns the hex color codes
for the specified palette.
}
\examples{
# Get a palette (when palettes are defined)
# vibe_palette("palette_name")
# vibe_palette("palette_name", n = 10)
}
