% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibeColors-package.R
\docType{package}
\name{vibeColors-package}
\alias{vibeColors}
\alias{vibeColors-package}
\title{vibeColors: Color Palettes for ggplot2}
\description{
A collection of color palettes designed for use with ggplot2. Provides discrete and continuous color scales that integrate seamlessly with the ggplot2 plotting system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yourusername.github.io/vibeColors}
  \item Report bugs at \url{https://github.com/yourusername/vibeColors/issues}
}

}
\author{
\strong{Maintainer}: Zepeng Mu \email{zmu@broadinstitute.org}

}
\keyword{internal}
